+++ qbittorrentapi/transfer.py
@@ -1,10 +1,11 @@
+import six
+
+from qbittorrentapi._version_support import v
 from qbittorrentapi.app import AppAPIMixIn
 from qbittorrentapi.decorators import alias
 from qbittorrentapi.decorators import aliased
 from qbittorrentapi.decorators import endpoint_introduced
 from qbittorrentapi.decorators import login_required
-from qbittorrentapi.decorators import response_json
-from qbittorrentapi.decorators import response_text
 from qbittorrentapi.definitions import APINames
 from qbittorrentapi.definitions import ClientCache
 from qbittorrentapi.definitions import Dictionary
@@ -48,20 +49,20 @@ class Transfer(ClientCache):
     @speedLimitsMode.setter
     def speedLimitsMode(self, v):
         """Implements
-        :meth:`~TransferAPIMixIn.transfer_toggle_speed_limits_mode`"""
+        :meth:`~TransferAPIMixIn.transfer_set_speed_limits_mode`"""
         self.speed_limits_mode = v
 
     @speed_limits_mode.setter
     def speed_limits_mode(self, v):
         """Implements
-        :meth:`~TransferAPIMixIn.transfer_toggle_speed_limits_mode`"""
-        self.toggle_speed_limits_mode(intended_state=v)
+        :meth:`~TransferAPIMixIn.transfer_set_speed_limits_mode`"""
+        self.set_speed_limits_mode(intended_state=v)
 
-    @alias("toggleSpeedLimitsMode")
-    def toggle_speed_limits_mode(self, intended_state=None, **kwargs):
+    @alias("setSpeedLimitsMode", "toggleSpeedLimitsMode", "toggle_speed_limits_mode")
+    def set_speed_limits_mode(self, intended_state=None, **kwargs):
         """Implements
-        :meth:`~TransferAPIMixIn.transfer_toggle_speed_limits_mode`"""
-        return self._client.transfer_toggle_speed_limits_mode(
+        :meth:`~TransferAPIMixIn.transfer_set_speed_limits_mode`"""
+        return self._client.transfer_set_speed_limits_mode(
             intended_state=intended_state, **kwargs
         )
 
@@ -139,7 +140,6 @@ class TransferAPIMixIn(AppAPIMixIn):
             self._transfer = Transfer(client=self)
         return self._transfer
 
-    @response_json(TransferInfoDictionary)
     @login_required
     def transfer_info(self, **kwargs):
         """
@@ -147,10 +147,14 @@ class TransferAPIMixIn(AppAPIMixIn):
 
         :return: :class:`TransferInfoDictionary` - `<https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-global-transfer-info>`_
         """  # noqa: E501
-        return self._get(_name=APINames.Transfer, _method="info", **kwargs)
+        return self._get(
+            _name=APINames.Transfer,
+            _method="info",
+            response_class=TransferInfoDictionary,
+            **kwargs
+        )
 
     @alias("transfer_speedLimitsMode")
-    @response_text(str)
     @login_required
     def transfer_speed_limits_mode(self, **kwargs):
         """
@@ -158,11 +162,20 @@ class TransferAPIMixIn(AppAPIMixIn):
 
         :return: ``1`` if alternative speed limits are currently enabled, ``0`` otherwise
         """
-        return self._get(_name=APINames.Transfer, _method="speedLimitsMode", **kwargs)
+        return self._get(
+            _name=APINames.Transfer,
+            _method="speedLimitsMode",
+            response_class=six.text_type,
+            **kwargs
+        )
 
-    @alias("transfer_toggleSpeedLimitsMode")
+    @alias(
+        "transfer_setSpeedLimitsMode",
+        "transfer_toggleSpeedLimitsMode",
+        "transfer_toggle_speed_limits_mode",
+    )
     @login_required
-    def transfer_toggle_speed_limits_mode(self, intended_state=None, **kwargs):
+    def transfer_set_speed_limits_mode(self, intended_state=None, **kwargs):
         """
         Sets whether alternative speed limits are enabled.
 
@@ -170,14 +183,26 @@ class TransferAPIMixIn(AppAPIMixIn):
                                Leaving None will toggle the current state.
         :return: None
         """
-        is_enabled = lambda: self.transfer_speed_limits_mode() == "1"  # noqa: E731
-        if intended_state is None or is_enabled() is not intended_state:
+        if intended_state is None:
             self._post(
                 _name=APINames.Transfer, _method="toggleSpeedLimitsMode", **kwargs
             )
+        elif v(self.app_web_api_version()) < v("2.8.14") and (
+            (self.transfer.speed_limits_mode == "1") is not bool(intended_state)
+        ):
+            self._post(
+                _name=APINames.Transfer, _method="toggleSpeedLimitsMode", **kwargs
+            )
+        else:
+            data = {"mode": 1 if intended_state else 0}
+            self._post(
+                _name=APINames.Transfer,
+                _method="setSpeedLimitsMode",
+                data=data,
+                **kwargs
+            )
 
     @alias("transfer_downloadLimit")
-    @response_text(int)
     @login_required
     def transfer_download_limit(self, **kwargs):
         """
@@ -185,10 +210,14 @@ class TransferAPIMixIn(AppAPIMixIn):
 
         :return: integer
         """
-        return self._get(_name=APINames.Transfer, _method="downloadLimit", **kwargs)
+        return self._get(
+            _name=APINames.Transfer,
+            _method="downloadLimit",
+            response_class=int,
+            **kwargs
+        )
 
     @alias("transfer_uploadLimit")
-    @response_text(int)
     @login_required
     def transfer_upload_limit(self, **kwargs):
         """
@@ -196,7 +225,9 @@ class TransferAPIMixIn(AppAPIMixIn):
 
         :return: integer
         """
-        return self._get(_name=APINames.Transfer, _method="uploadLimit", **kwargs)
+        return self._get(
+            _name=APINames.Transfer, _method="uploadLimit", response_class=int, **kwargs
+        )
 
     @alias("transfer_setDownloadLimit")
     @login_required